Remove the GameContext association from Entity
It makes sense for the GameContext to know about every entity, but each
entity shouldn't persist the GameContext for all time.

This makes it tricky to build entities as you would previously be forced
to construct a new game, just to test entities.

The only purpose for the association was to do things such as querying
the current position. In these cases, we now pass the context in
explicitly.