Fix a bug in the A* implementation
The G score was being incorrectly calculated. G scores would not
accumulate - every time we would generate a completely new gscore value,
rather than adding onto the existing one from the previous iteration.

When looking for the next cheapest node in each A* iteration, this could
leave multiple nodes with the same fScore value, even though one could
be clearly closer to the goal than the other. In this case, if there are
multiple nodes with the same score, then the first matching one is
chosen.

This would sometimes cause the pathfinding library to take a suboptimal
path to the destination, but still, to the correct destination.

Thanks to Josh for finding this in his external tests.